% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opusminer.R
\name{opus}
\alias{opus}
\title{OPUS Miner: Filtered Top-k Association Discovery of Self-Sufficient Itemsets}
\usage{
opus(transactions = NULL, k = 100, format = "dataframe", sep = " ", ...)
}
\arguments{
\item{transactions}{A filename, list, or object of class
\code{\link[arules]{transactions}} (\code{arules}).}

\item{k}{The number of itemsets to return, an integer (default 100).}

\item{format}{The output format ("dataframe", default, or "itemsets").}

\item{sep}{The separator between items (for files, default " ").}

\item{...}{Optional additional parameters (see details).}
}
\value{
The top k productive, non-redundant itemsets, with relevant
  statistics, in the form of a data frame, object of class
  \code{\link[arules]{itemsets}} (\code{arules}), or a list.
}
\description{
\code{opus} finds the top k productive, non-redundant itemsets on the
measure of interest (leverage or lift) using the OPUS Miner algorithm.
}
\details{
\code{opus} provides an interface to the OPUS Miner algorithm (implemented in
C++) to find the top k productive, non-redundant itemsets by leverage
(default) or lift.

\code{transactions} should be a filename, list (of transactions, each list
element being a vector of character values representing item labels), or an
object of class \code{\link[arules]{transactions}} (\code{arules}).

Files should be in the format of a list of transactions, one line per
transaction, each transaction (ie, line) being a sequence of item labels,
separated by the character specified by the parameter \code{sep} (default
" ").  (Alternatively, files can be read seaparately using the \code{\link{read_transactions}}
function.)

\code{format} should be specified as either "dataframe" (the default) or
"itemsets", and any other value will return a list.

The optional additional parameters are as follows:
\itemize{
  \item \code{print_closures}
        whether to also return the closure for each itemset (default FALSE)
  \item \code{filter_itemsets}
        whether to filter itemsets that are not independently productive
        (default TRUE)
  \item \code{search_by_lift}
        make lift (rather than leverage) the measure of interest (default
        FALSE)
  \item \code{correct_for_mult_compare}
        whether to correct alpha for the size of the search space (default
        TRUE)
  \item \code{redundancy_tests}
        whether to allow redundant itemsets (default TRUE)
}
}
\references{
Webb, G.I. & Vreeken, J. (2014) Efficient Discovery of the Most Interesting
Associations. ACM Transactions on Knowledge Discovery from Data. 8(3), Art.
no. 15.
}
